varios cambios

+++ CryptoTrackerApp/Classes/FavoriteCryptos.cs
@@ -17,7 +17,7 @@ namespace CryptoTrackerApp.Classes
         public string[] IdCrypto { get; set; }
 
         [Column("Umbral")]
-        public int[] Umbral { get; set; }
+        public float[] Umbral { get; set; }
     }
